refactor(core): keep the #609 helpers inside the guardrail ratchets
Move the sse_stream_integrity tests into their own file so the panic
ratchet does not count test .expect() calls as production panics, and
use from_utf8_lossy on the already-validated prefix instead of
unwrap_or_default so the swallowed-error budget stays flat.

No behavior change; scripts/check_guardrails.sh --skip-slow now passes.
